// src/_config/payload-generated-schema.ts
export const enum_pathways_flow = pgEnum('enum_pathways_flow', ['wip', 'submitted', 'published'])
// ...
export const enum_users_status = pgEnum('enum_users_status', [
  'pending',
  'active',
  'hiatus',
  'inactive',
])
// ...
export const countries = pgTable(
  'countries',
  {
    id: varchar('id').primaryKey(),
    flow: enum_countries_flow('flow').notNull().default('wip'),
    currentLifecycle_published: integer('current_lifecycle_published_id').references(
      (): AnyPgColumn => countries_record.id,
      {
        onDelete: 'set null',
      }
    ),
    currentLifecycle_locked: boolean('current_lifecycle_locked').default(false),
    currentLifecycle_changed_at: timestamp('current_lifecycle_changed_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    currentLifecycle_changed_by: integer('current_lifecycle_changed_by_id').references(
      () => users.id,
      {
        onDelete: 'set null',
      }
    ),
    unMember: boolean('un_member').default(false),
    name: varchar('name').notNull(),
    currency: jsonb('currency').default(sql`'[]'::jsonb`),
    language: jsonb('language').default(sql`'[]'::jsonb`),
    countryImage: integer('country_image_id').references((): AnyPgColumn => country_images.id, {
      onDelete: 'set null',
    }),
    mapSvgPath: varchar('map_svg_path'),
    'iso-2': varchar('iso_2'),
  },
  (columns) => [
    index('countries_current_lifecycle_current_lifecycle_published_idx').on(
      columns.currentLifecycle_published
    ),
    index('countries_current_lifecycle_changed_current_lifecycle_ch_idx').on(
      columns.currentLifecycle_changed_by
    ),
    uniqueIndex('countries_name_idx').on(columns.name),
    index('countries_country_image_idx').on(columns.countryImage),
  ]
)
// ...
export const pathways = pgTable(
  'pathways',
  {
    id: serial('id').primaryKey(),
    flow: enum_pathways_flow('flow').notNull().default('wip'),
    currentLifecycle_published: integer('current_lifecycle_published_id').references(
      () => pathways_record.id,
      {
        onDelete: 'set null',
      }
    ),
    currentLifecycle_locked: boolean('current_lifecycle_locked').default(false),
    currentLifecycle_changed_at: timestamp('current_lifecycle_changed_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    currentLifecycle_changed_by: integer('current_lifecycle_changed_by_id').references(
      () => users.id,
      {
        onDelete: 'set null',
      }
    ),
    country: varchar('country_id')
      .notNull()
      .references(() => countries.id, {
        onDelete: 'set null',
      }),
    name: varchar('name').notNull(),
    commonName: varchar('common_name'),
    link: varchar('link').notNull(),
    description: varchar('description').notNull(),
    currency: varchar('currency'),
    costData: jsonb('cost_data').default(sql`'{"min":0,"max":0,"na":false}'::jsonb`),
    processingTime: jsonb('processing_time').default(
      sql`'{"min":{"qty":0,"uom":null,"business":false},"max":{"qty":0,"uom":null,"business":false}}'::jsonb`
    ),
    visaDuration: jsonb('visa_duration').default(
      sql`'{"min":{"qty":0,"uom":null,"business":false},"max":{"qty":0,"uom":null,"business":false}}'::jsonb`
    ),
    renewable: boolean('renewable'),
    citizenshipPathway: boolean('citizenship_pathway'),
    residencyPathway: boolean('residency_pathway'),
    reunificationPathway: boolean('reunification_pathway'),
    pipelineNotesRenewable: varchar('pipeline_notes_renewable'),
    pipelineNotesCitizenshipPathway: varchar('pipeline_notes_citizenship_pathway'),
    pipelineNotesResidencyPathway: varchar('pipeline_notes_residency_pathway'),
    pipelineNotesReunificationPathway: varchar('pipeline_notes_reunification_pathway'),
    flags: jsonb('flags').default(sql`'[]'::jsonb`),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    deletedAt: timestamp('deleted_at', { mode: 'string', withTimezone: true, precision: 3 }),
  },
  (columns) => [
    index('pathways_current_lifecycle_current_lifecycle_published_idx').on(
      columns.currentLifecycle_published
    ),
    index('pathways_current_lifecycle_changed_current_lifecycle_cha_idx').on(
      columns.currentLifecycle_changed_by
    ),
    index('pathways_country_idx').on(columns.country),
    index('pathways_updated_at_idx').on(columns.updatedAt),
    index('pathways_created_at_idx').on(columns.createdAt),
    index('pathways_deleted_at_idx').on(columns.deletedAt),
  ]
)

export const pathways_rels = pgTable(
  'pathways_rels',
  {
    id: serial('id').primaryKey(),
    order: integer('order'),
    parent: integer('parent_id').notNull(),
    path: varchar('path').notNull(),
    'pathways-recordID': integer('pathways_record_id'),
    usersID: integer('users_id'),
    'pathway-categoriesID': integer('pathway_categories_id'),
  },
  (columns) => [
    index('pathways_rels_order_idx').on(columns.order),
    index('pathways_rels_parent_idx').on(columns.parent),
    index('pathways_rels_path_idx').on(columns.path),
    index('pathways_rels_pathways_record_id_idx').on(columns['pathways-recordID']),
    index('pathways_rels_users_id_idx').on(columns.usersID),
    index('pathways_rels_pathway_categories_id_idx').on(columns['pathway-categoriesID']),
    foreignKey({
      columns: [columns['parent']],
      foreignColumns: [pathways.id],
      name: 'pathways_rels_parent_fk',
    }).onDelete('cascade'),
    foreignKey({
      columns: [columns['pathways-recordID']],
      foreignColumns: [pathways_record.id],
      name: 'pathways_rels_pathways_record_fk',
    }).onDelete('cascade'),
    foreignKey({
      columns: [columns['usersID']],
      foreignColumns: [users.id],
      name: 'pathways_rels_users_fk',
    }).onDelete('cascade'),
    foreignKey({
      columns: [columns['pathway-categoriesID']],
      foreignColumns: [pathway_categories.id],
      name: 'pathways_rels_pathway_categories_fk',
    }).onDelete('cascade'),
  ]
)
// ...
export const pathways_record = pgTable(
  'pathways_record',
  {
    id: serial('id').primaryKey(),
    title: varchar('title'),
    baseId: numeric('base_id', { mode: 'number' }).notNull(),
    change_at: timestamp('change_at', { mode: 'string', withTimezone: true, precision: 3 }),
    change_by: integer('change_by_id').references(() => users.id, {
      onDelete: 'set null',
    }),
    change_system: boolean('change_system').default(false),
    transition_at: timestamp('transition_at', { mode: 'string', withTimezone: true, precision: 3 }),
    transition_by: integer('transition_by_id').references(() => users.id, {
      onDelete: 'set null',
    }),
    transition_system: boolean('transition_system').default(false),
    transition_from: varchar('transition_from'),
    transition_to: varchar('transition_to'),
    transition_actionKey: varchar('transition_action_key'),
    snapshot_country: varchar('snapshot_country_id').references(() => countries.id, {
      onDelete: 'set null',
    }),
    snapshot_name: varchar('snapshot_name'),
    snapshot_commonName: varchar('snapshot_common_name'),
    snapshot_link: varchar('snapshot_link'),
    snapshot_description: varchar('snapshot_description'),
    snapshot_currency: varchar('snapshot_currency'),
    snapshot_costData: jsonb('snapshot_cost_data'),
    snapshot_processingTime: jsonb('snapshot_processing_time'),
    snapshot_visaDuration: jsonb('snapshot_visa_duration'),
    snapshot_renewable: boolean('snapshot_renewable'),
    snapshot_citizenshipPathway: boolean('snapshot_citizenship_pathway'),
    snapshot_residencyPathway: boolean('snapshot_residency_pathway'),
    snapshot_reunificationPathway: boolean('snapshot_reunification_pathway'),
    snapshot_pipelineNotesRenewable: varchar('snapshot_pipeline_notes_renewable'),
    snapshot_pipelineNotesCitizenshipPathway: varchar(
      'snapshot_pipeline_notes_citizenship_pathway'
    ),
    snapshot_pipelineNotesResidencyPathway: varchar('snapshot_pipeline_notes_residency_pathway'),
    snapshot_pipelineNotesReunificationPathway: varchar(
      'snapshot_pipeline_notes_reunification_pathway'
    ),
    snapshot_flags: jsonb('snapshot_flags'),
    _options: jsonb('_options').default(sql`'{"flow":[]}'::jsonb`),
  },
  (columns) => [
    index('pathways_record_title_idx').on(columns.title),
    index('pathways_record_base_id_idx').on(columns.baseId),
    index('pathways_record_change_change_at_idx').on(columns.change_at),
    index('pathways_record_change_change_by_idx').on(columns.change_by),
    index('pathways_record_transition_transition_at_idx').on(columns.transition_at),
    index('pathways_record_transition_transition_by_idx').on(columns.transition_by),
    index('pathways_record_transition_transition_from_idx').on(columns.transition_from),
    index('pathways_record_transition_transition_to_idx').on(columns.transition_to),
    index('pathways_record_transition_transition_action_key_idx').on(columns.transition_actionKey),
    index('pathways_record_snapshot_snapshot_country_idx').on(columns.snapshot_country),
  ]
)
// ...
export const users = pgTable(
  'users',
  {
    id: serial('id').primaryKey(),
    name: varchar('name').notNull(),
    firstName: varchar('first_name'),
    lastName: varchar('last_name'),
    status: enum_users_status('status').notNull().default('pending'),
    lastLogin: timestamp('last_login', { mode: 'string', withTimezone: true, precision: 3 }),
    weeklyDays: jsonb('weekly_days').default(
      sql`'{"M":{"morning":false,"afternoon":false,"evening":false},"T":{"morning":false,"afternoon":false,"evening":false},"W":{"morning":false,"afternoon":false,"evening":false},"Th":{"morning":false,"afternoon":false,"evening":false},"F":{"morning":false,"afternoon":false,"evening":false},"S":{"morning":false,"afternoon":false,"evening":false},"Su":{"morning":false,"afternoon":false,"evening":false}}'::jsonb`
    ),
    timezone: enum_users_timezone('timezone'),
    weeklyHours: enum_users_weekly_hours('weekly_hours'),
    isDirector: boolean('is_director').default(false),
    isHead: boolean('is_head').default(false),
    volunteerAgreement: boolean('volunteer_agreement').default(false),
    volunteerAgreementSignature: varchar('volunteer_agreement_signature'),
    volunteerAgreementSignedOn: timestamp('volunteer_agreement_signed_on', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    addedToDiscord: boolean('added_to_discord').default(false),
    discordHandle: varchar('discord_handle'),
    folder: integer('folder_id').references(() => payload_folders.id, {
      onDelete: 'set null',
    }),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    deletedAt: timestamp('deleted_at', { mode: 'string', withTimezone: true, precision: 3 }),
    email: varchar('email'),
    username: varchar('username'),
    resetPasswordToken: varchar('reset_password_token'),
    resetPasswordExpiration: timestamp('reset_password_expiration', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    salt: varchar('salt'),
    hash: varchar('hash'),
    loginAttempts: numeric('login_attempts', { mode: 'number' }).default(0),
    lockUntil: timestamp('lock_until', { mode: 'string', withTimezone: true, precision: 3 }),
  },
  (columns) => [
    index('users_folder_idx').on(columns.folder),
    index('users_updated_at_idx').on(columns.updatedAt),
    index('users_created_at_idx').on(columns.createdAt),
    index('users_deleted_at_idx').on(columns.deletedAt),
    uniqueIndex('users_username_idx').on(columns.username),
  ]
)